resolve problem can't load assembly in first time open button
