+++ ConnectorNavisworks/ConnectorNavisworks/Entry/SpeckleNavisworksCommand.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using System.Windows.Forms;
 using System.Windows.Forms.Integration;
@@ -34,8 +35,9 @@ namespace Speckle.ConnectorNavisworks.Entry
   {
     public override Control CreateControlPane()
     {
+      
       Setup.Init(ConnectorBindingsNavisworks.HostAppNameVersion, ConnectorBindingsNavisworks.HostAppName);
-      AppDomain.CurrentDomain.AssemblyResolve += OnAssemblyResolve;
+      
       try
       {
         InitAvalonia();
@@ -91,18 +93,6 @@ namespace Speckle.ConnectorNavisworks.Entry
       BuildAvaloniaApp().SetupWithoutStarting();
     }
 
-    private static Assembly OnAssemblyResolve(object sender, ResolveEventArgs args)
-    {
-      Assembly a = null;
-      var name = args.Name.Split(',')[0];
-      var path = Path.GetDirectoryName(typeof(RibbonHandler).Assembly.Location);
-
-      var assemblyFile = Path.Combine(path ?? string.Empty, name + ".dll");
-
-      if (File.Exists(assemblyFile))
-        a = Assembly.LoadFrom(assemblyFile);
-
-      return a;
-    }
+    
   }
 }
