+++ ConnectorNavisworks/ConnectorNavisworks2023/ConnectorNavisworks2023.csproj
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <!--EXTERNAL_PROPERTIES: AppData-->
-<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+<Project Sdk="Microsoft.NET.Sdk">
 
 	<!-- Project Config -->
 	<PropertyGroup>
